+++ src/game/time.c
@@ -6,7 +6,7 @@
 /*   By: mkling <mkling@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/09 00:20:56 by alex              #+#    #+#             */
-/*   Updated: 2025/03/20 18:58:52 by mkling           ###   ########.fr       */
+/*   Updated: 2025/03/21 14:58:03 by mkling           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ void	update_delta_time(t_cub *cub)
 	current_frame = get_microseconds();
 	delta_ms = current_frame - cub->display.last_frame;
 	cub->display.last_frame = current_frame;
-	cub->display.delta_time = ((double)delta_ms) / 1000000.00;
+	cub->display.delta_time = ((float)delta_ms) / 1000000.00;
 	cub->display.counter += cub->display.delta_time;
 	if (cub->display.counter > 1)
 	{
