+++ src/transform/rotate.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   rotate.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: vgodoy <vgodoy@student.42.fr>              +#+  +:+       +#+        */
+/*   By: mkling <mkling@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/04 12:02:27 by mkling            #+#    #+#             */
-/*   Updated: 2025/03/12 16:35:37 by vgodoy           ###   ########.fr       */
+/*   Updated: 2025/03/21 14:58:03 by mkling           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 t_pt	normalize_vector(t_pt *point)
 {
 	t_pt	unit_vector;
-	double	magnitude;
+	float	magnitude;
 
 	unit_vector = *point;
 	magnitude = sqrt(point->x * point->x + point->y * point->y);
@@ -26,11 +26,11 @@ t_pt	normalize_vector(t_pt *point)
 	return (unit_vector);
 }
 
-void	rotate_point(t_cub *cub, t_pt *point, t_pt *center, double angle)
+void	rotate_point(t_cub *cub, t_pt *point, t_pt *center, float angle)
 {
-	double	radian;
-	double	cos_angle;
-	double	sin_angle;
+	float	radian;
+	float	cos_angle;
+	float	sin_angle;
 	t_pt	centered;
 	t_pt	rotated;
